rename veterans-personal-info directory
